Configured Dnsmasq as a DHCP and DNS server
Summary of changes:
1. Added gateway VM
2. Installed Docker on gateway VM
3. Build Dnsmasq Docker image only if necessary
4. configure-gateway-network.sh script is now idempotent
5. Do not write to stdout when checking docker binary
6. Start DNSmasq container only if necessary
7. Assign hostnames, ip addresses and domain via Dnsmasq
8. Keep only the default route via our gateway
9. Configure network interfaces using NetworkManager

Each VM has an additional network interface that connects to the
same Virtualbox internal network. All traffic is then routed by
default through this network interface and the default gateway.
With this change we ensure that all our traffic is isolated and
independent from external actors (only the hosts attached to the
Virtualbox internal network can see each other).
Note that we added a script that is executed on all "network
interface up" events and sets the default route through
our gateway. This was necessary because Vagrant configures a
NATed network interface by design (and this feature cannot be
disabled and is necessary for Vagrant to work). Normally there
should be no need to setup this script in an environment not
managed by Vagrant.

The network interface is configured via NetworkManager with the
exception of the Gateway VM that is configured using
/etc/network/interfaces (to avoid setup complications). We chose
NetworkManager because we are currently working in a mixed
(Ubuntu and CentOS) environment, and NetworkManager is available
on both (and is the default on CentOS, while it has to be
installed on Ubuntu server).

Each VM is assigned an IPv4 address, a hostname and a domain via
DHCP. IPv4 reservation is configured in DNSmasq using each VM
network interface MAC address.

Moreover each VM is reachable from <vm-name>.scissor-project.com
and simply from <vm-name>. Dnsmasq was configured to handle both
scenarios.
Examples:
- ping kafka
- ping kafka.scissor-project.com

Note that to change the domain (now set to scissor-project.com)
you need to edit both the Vagrantfile and the Dnsmasq
configuration file (dnsmasq.conf) and possibly rebuild the related
Docker image, stop all the instances of Dnsmasq based on the old
image and start a new instance based on the freshly built image.